+++ Source/SwiftLintBuiltInRules/Rules/Lint/UnneededThrowsRuleExamples.swift
@@ -1,0 +1,258 @@
+internal struct UnneededThrowsRuleExamples {
+    static let nonTriggeringExamples = [
+        Example("""
+            func foo() throws {
+                try bar()
+            }
+            """),
+        Example("""
+            func foo() throws {
+                throw Example.failure
+            }
+            """),
+        Example("""
+            func foo(_ bar: () throws -> T) rethrows -> Int {
+                try items.map { try bar() }
+            }
+            """),
+        Example("""
+            func foo() {
+                func bar() throws {
+                    try baz()
+                }
+                try? bar()
+            }
+            """),
+        Example("""
+            protocol Foo {
+                func bar() throws
+            }
+            """),
+        Example("""
+            func foo() throws {
+                guard false else {
+                    throw Example.failure
+                }
+            }
+            """),
+        Example("""
+            func foo() throws {
+                do { try bar() } 
+                catch {
+                    throw Example.failure
+                }
+            }
+            """),
+        Example("""
+            func foo() throws {
+                do { try bar() } 
+                catch {
+                    try baz()
+                }
+            }
+            """),
+        Example("""
+        func foo() throws {
+            do {
+                throw Example.failure
+            } catch {
+                do {
+                    throw Example.failure
+                } catch {
+                    throw Example.failure
+                }
+            }
+        }
+        """),
+        Example("""
+            func foo() throws {
+                switch bar {
+                case 1: break
+                default: try bar()
+                }
+            }
+            """),
+        Example("""
+            var foo: Int {
+                get throws {
+                    try bar
+                }
+            }
+            """),
+    ]
+
+    static let triggeringExamples = [
+        Example("func foo() ↓throws {}"),
+        Example("let foo: () ↓throws -> Void = {}"),
+        Example("let foo: (() ↓throws -> Void)? = {}"),
+        Example("func foo(bar: () throws -> Void) ↓rethrows {}"),
+        Example("init() ↓throws {}"),
+        Example("""
+            func foo() ↓throws {
+                bar()
+            }
+            """),
+        Example("""
+            func foo() {
+                func bar() ↓throws {}
+                bar()
+            }
+            """),
+        Example("""
+            func foo() {
+                func bar() ↓throws {
+                    baz()
+                }
+                bar()
+            }
+            """),
+        Example("""
+            func foo() {
+                func bar() ↓throws {
+                    baz()
+                }
+                try? bar()
+            }
+            """),
+        Example("""
+            func foo() ↓throws {
+                func bar() ↓throws {
+                    baz()
+                }
+            }
+            """),
+        Example("""
+            func foo() ↓throws {
+                do { try bar() } 
+                catch {}
+            }
+            """),
+        Example("""
+            func foo() ↓throws {
+                do {} 
+                catch {}
+            }
+            """),
+        Example("""
+            func foo() {
+                do {
+                    func bar() ↓throws {}
+                    try bar()
+                } catch {}
+            }
+            """),
+        Example("""
+            func foo() ↓throws {
+                do {
+                    try bar()
+                    func baz() throws { try bar() }
+                    try baz()
+                } catch {}
+            }
+            """),
+        Example("""
+            func foo() ↓throws {
+                do {
+                    try bar()
+                } catch {
+                    do {
+                        throw Example.failure
+                    } catch {}
+                }
+            }
+            """),
+        Example("""
+            func foo() ↓throws {
+                do {
+                    try bar()
+                } catch {
+                    do {
+                        try bar()
+                        func baz() ↓throws {}
+                        try baz()
+                    } catch {}
+                }
+            }
+            """),
+        Example("""
+            func foo() ↓throws {
+                switch bar {
+                case 1: break
+                default: break
+                }
+            }
+            """),
+        Example("""
+            func foo() ↓throws {
+                _ = try? bar()
+            }
+            """),
+        Example("""
+            func foo() ↓throws {
+                Task {
+                    try bar()
+                }
+            }
+            """),
+        Example("""
+            func foo() throws {
+                try bar()
+                Task {
+                    func baz() ↓throws {}
+                }
+            }
+            """),
+        Example("""
+            var foo: Int {
+                get ↓throws {
+                    0
+                }
+            }
+            """),
+    ]
+
+    static let corrections = [
+        Example("func foo() ↓throws {}"): Example("func foo() {}"),
+        Example("init() ↓throws {}"): Example("init() {}"),
+        Example("""
+        func foo() {
+            func bar() ↓throws {}
+            bar()
+        }
+        """): Example("""
+            func foo() {
+                func bar() {}
+                bar()
+            }
+            """),
+        Example("""
+            var foo: Int {
+                get ↓throws {
+                    0
+                }
+            }
+            """): Example("""
+            var foo: Int {
+                get {
+                    0
+                }
+            }
+            """),
+        Example("""
+            let foo: () ↓throws -> Void = {}
+            """): Example("""
+            let foo: () -> Void = {}
+            """),
+        Example("""
+            func foo() ↓throws {
+                do {}
+                catch {}
+            }
+            """): Example("""
+            func foo() {
+                do {}
+                catch {}
+            }
+            """),
+    ]
+}